Update to ensure_valid_license_family so that existing recipes don't break
